prompt/ai-customer-service/requirements.txt
python-dotenv==1.1.0
openai==1.73.0
